requirements.txt
flask>=2.0.0
werkzeug>=2.0.0
PyMuPDF>=1.21.0
openai>=1.0.0
llama-index>=0.8.0
python-dotenv>=0.19.0
numpy>=1.21.0
pandas>=1.3.0
PyPDF2>=3.0.0
python-magic-bin>=0.4.14
tabula-py>=2.7.0
openpyxl>=3.0.0
llama-index-llms-openai